sched/walt: Improve the scheduler
This change is for general scheduler improvement.

+++ include/linux/sched.h
@@ -825,7 +825,7 @@ struct task_struct {
 	struct list_head grp_list;
 	u64 cpu_cycles;
 	bool misfit;
-	u8 unfilter;
+	u32 unfilter;
 #endif
 
 #ifdef CONFIG_CGROUP_SCHED
